+++ src/main/java/Duke.java
@@ -113,84 +113,3 @@ public class Duke {
         }
     }
 }
-
-class Task {
-    String description;
-    boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public void markDone() {
-        this.isDone = true;
-    }
-
-    public String toString() {
-        if (this.isDone) {
-            return "[✓] " + this.description;
-        } else {
-            return "[✗] " + this.description;
-        }
-    }
-}
-
-class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    public String toString() {
-        if (this.isDone) {
-            return "[T][✓] " + this.description;
-        } else {
-            return "[T][✗] " + this.description;
-        }
-    }
-}
-
-class Deadline extends Task {
-    public String date;
-
-    public Deadline(String description) {
-        super(description);
-        String[] arr = description.split(" /by ");
-        this.description = arr[0];
-        this.date = arr[1];
-    }
-
-    public String toString() {
-        if (this.isDone) {
-            return "[D][✓] " + this.description + " (by: " + this.date + ")";
-        } else {
-            return "[D][✗] " + this.description + " (by: " + this.date + ")";
-        }
-    }
-}
-
-class Event extends Task {
-    public String date;
-
-    public Event(String description) {
-        super(description);
-        String[] arr = description.split(" /at ");
-        this.description = arr[0];
-        this.date = arr[1];
-    }
-
-    public String toString() {
-        if (this.isDone) {
-            return "[E][✓] " + this.description + " (at: " + this.date + ")";
-        } else {
-            return "[E][✗] " + this.description + " (at: " + this.date + ")";
-        }
-    }
-}
-
-class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
